Change calculator file.
so i was really stuck with the logic of the "program". Couldnt find the
solution so i made a fresh start.
The idea was to keep the values of the numbers and operators in
separated functions and then work with the functions.
Its not finished yet but it does the core work.